wgan_requirements.txt
torch
torchvision
matplotlib
numpy
cma
Pillow